api/requirements.txt
# FastAPI Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography==43.0.3

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Validation & Serialization
pydantic==2.10.0
pydantic-settings==2.6.0
email-validator==2.2.0

# AI Gateway
openai==1.55.0
anthropic==0.39.0
langchain==0.3.7
langchain-openai==0.2.9
langchain-anthropic==0.3.0

# Utils
python-dotenv==1.0.1
httpx==0.28.0
aiofiles==24.1.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.0
